backend/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
b2sdk==2.5.1
bidict==0.23.1
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
decorator==4.4.2
Django==5.1.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-rest-framework==0.1.0
django-sslserver==0.22
djangorestframework==3.15.2
dnspython==2.6.1
elevenlabs==1.9.0
eventlet==0.37.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.4.1
greenlet==3.1.1
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
hyperlink==21.0.0
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
logfury==1.0.1
MarkupSafe==2.1.5
moviepy==1.0.3
numpy==2.1.1
pillow==10.4.0
proglog==0.1.10
prompt_toolkit==3.0.48
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.2
redis==5.1.1
requests==2.32.3
scipy==1.14.1
service-identity==24.1.0
setuptools==74.1.2
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.4
wsproto==1.2.0
zope.interface==7.1.1
